Phase 2: Extract theme system into theme-manager.js module
Extracted complete theme system into standalone ES6 module:

Files Created:
- public/theme-manager.js (714 lines)
  - All 7 theme definitions (modern, digital, dark, arctic, sunset, neon, vintage)
  - Theme application with CSS variable updates
  - Three.js scene color update integration
  - localStorage persistence
  - Public API: applyTheme, updateSceneColors, getCurrentTheme, initializeTheme, getThemes, getTheme

Files Modified:
- public/app.js (11,699 lines, down from ~12,382)
  - Removed 683 lines of theme code (lines 73-755)
  - Added imports from theme-manager module
  - Added wrapper functions for scene/CONFIG integration
  - Updated theme card click handler to use applyThemeWithScene()

- public/index.html
  - Added theme-manager.js module loading before app.js

- 5-REFACTORING_LOG.md
  - Updated Phase 2 status to complete
  - Documented all changes and metrics

Module Architecture:
- ES6 modules with clean imports/exports
- Script loading order: config.js → constants.js → theme-manager.js → aircraft-svg-system.js → app.js
- Backward compatible - all theme functions work via imports
- Self-contained module with no external dependencies (except SafeStorage)

Metrics:
- Lines extracted: 683
- app.js reduction: 5.5%
- Phase target: ~755 lines (achieved: 683 lines, 90.5% of target)

Status: Complete, pending user validation of theme switching functionality
